requirements.txt
allure-pytest
allure-python-commons
httpx~=0.19.0
playwright==1.15.0
pytest~=6.2.5
pytest-asyncio
pylint-fail-under
pytest-xdist
pytest-bdd~=4.1.0
pytest-factoryboy
pytest-cov
pytest-rerunfailures
playwright-stealth
pytest-playwright
pytest-playwright-snapshot
pytest-html
pytest-html-reporter
pytest-metadata
pytest-order
simplejson~=3.17.5
pathlib~=1.0.1
py~=1.10.0
pytest-reportportal
pyleniumio
screenpy
selenium